+++ Common/src/main/java/com/tristankechlo/random_mob_sizes/config/EntityTypeList.java
@@ -2,6 +2,7 @@ package com.tristankechlo.random_mob_sizes.config;
 
 import com.google.common.collect.ImmutableList;
 import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
 import com.tristankechlo.random_mob_sizes.IPlatformHelper;
@@ -12,6 +13,7 @@ import net.minecraft.util.GsonHelper;
 import net.minecraft.world.entity.EntityType;
 
 import java.util.Arrays;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
 import java.util.function.Supplier;
@@ -24,7 +26,7 @@ public final class EntityTypeList implements Supplier<List<EntityType<?>>> {
     private final ImmutableList<String> defaultValue;
 
     public EntityTypeList(String key, List<String> stringValues) {
-        this(key, stringValues, parseList(stringValues, key));
+        this(key, stringValues, parseList(stringValues, key, () -> {}));
     }
 
     public EntityTypeList(String key, EntityType<?>... entityTypes) {
@@ -39,7 +41,7 @@ public final class EntityTypeList implements Supplier<List<EntityType<?>>> {
     }
 
     public void setToDefault() {
-        this.cachedValue = parseList(defaultValue, key);
+        this.cachedValue = parseList(defaultValue, key, () -> {});
         this.parsedValues = ImmutableList.copyOf(defaultValue);
     }
 
@@ -51,20 +53,23 @@ public final class EntityTypeList implements Supplier<List<EntityType<?>>> {
         json.add(key, array);
     }
 
-    public void deserialize(JsonObject json) {
+    public void deserialize(JsonObject json, Runnable setMakeBackup) {
         try {
             JsonArray array = GsonHelper.getAsJsonArray(json, key);
+            Iterator<JsonElement> elements = array.iterator();
             ImmutableList.Builder<String> builder = ImmutableList.builder();
-            array.forEach(element -> {
+            while (elements.hasNext()) {
+                JsonElement element = elements.next();
                 String value = GsonHelper.convertToString(element, key);
                 builder.add(value);
-            });
+            }
             this.parsedValues = builder.build();
-            this.cachedValue = parseList(parsedValues, key);
+            this.cachedValue = parseList(parsedValues, key, setMakeBackup);
         } catch (Exception e) {
             RandomMobSizes.LOGGER.error("Error while parsing config value '{}', using default value", key);
+            RandomMobSizes.LOGGER.error(e.getMessage());
             setToDefault();
-            throw new ConfigParseException(e.getMessage());
+            setMakeBackup.run();
         }
     }
 
@@ -73,7 +78,7 @@ public final class EntityTypeList implements Supplier<List<EntityType<?>>> {
         return cachedValue;
     }
 
-    private static ImmutableList<EntityType<?>> parseList(List<String> values, String key) {
+    private static ImmutableList<EntityType<?>> parseList(List<String> values, String key, Runnable setMakeBackup) {
         ImmutableList.Builder<EntityType<?>> builder = ImmutableList.builder();
         for (String value : values) {
             if (value.endsWith(":*")) {
@@ -81,6 +86,7 @@ public final class EntityTypeList implements Supplier<List<EntityType<?>>> {
                 String namespace = value.substring(0, value.length() - 2);
                 if (!IPlatformHelper.INSTANCE.isModLoaded(namespace)) {
                     RandomMobSizes.LOGGER.error("Skipping unknown wildcard: '{}' of config value '{}'", namespace, key);
+                    setMakeBackup.run();
                     continue;
                 }
                 Registry.ENTITY_TYPE.stream()
@@ -95,9 +101,11 @@ public final class EntityTypeList implements Supplier<List<EntityType<?>>> {
                         builder.add(type.get());
                     } else {
                         RandomMobSizes.LOGGER.error("Skipping disabled EntityType: '{}' of config value '{}'", value, key);
+                        setMakeBackup.run();
                     }
                 } else {
                     RandomMobSizes.LOGGER.error("Skipping unknown EntityType: '{}' of config value '{}'", value, key);
+                    setMakeBackup.run();
                 }
             }
         }