+++ Common/src/main/java/com/tristankechlo/random_mob_sizes/config/RandomMobSizesConfig.java
@@ -4,8 +4,9 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.tristankechlo.random_mob_sizes.RandomMobSizes;
 import com.tristankechlo.random_mob_sizes.sampler.AttributeScalingTypes;
-import com.tristankechlo.random_mob_sizes.sampler.GaussianScalingSampler;
 import com.tristankechlo.random_mob_sizes.sampler.ScalingSampler;
+import com.tristankechlo.random_mob_sizes.sampler.types.StaticScalingSampler;
+import net.minecraft.util.GsonHelper;
 import net.minecraft.world.entity.EntityType;
 
 import java.util.List;
@@ -15,47 +16,48 @@ public final class RandomMobSizesConfig {
     private static final BooleanValue KEEP_SCALING_ON_CONVERSION = new BooleanValue("keep_scaling_on_conversion", true);
     private static ScalingSampler defaultSampler = createDefaultSampler();
     private static final String DEFAULT_SAMPLER_NAME = "default_scaling";
-    private static final EntityTypeList WHITELIST = new EntityTypeList("whitelist", List.of("minecraft:*"));
-    private static final EntityTypeList BLACKLIST = new EntityTypeList("blacklist", EntityType.SHULKER, EntityType.WITHER);
+    private static final EntityTypeList INCLUDE_LIST = new EntityTypeList("include_list", List.of("minecraft:*"));
+    private static final EntityTypeList EXCLUDE_LIST = new EntityTypeList("exclude_list", EntityType.SHULKER, EntityType.WITHER);
     public static final ScalingOverrides SCALING_OVERRIDES = new ScalingOverrides();
 
     public static void setToDefault() {
         KEEP_SCALING_ON_CONVERSION.setToDefault();
         defaultSampler = createDefaultSampler();
-        WHITELIST.setToDefault();
-        BLACKLIST.setToDefault();
+        INCLUDE_LIST.setToDefault();
+        EXCLUDE_LIST.setToDefault();
         SCALING_OVERRIDES.setToDefault();
     }
 
     public static JsonObject serialize() {
         JsonObject json = new JsonObject();
         json.add(DEFAULT_SAMPLER_NAME, defaultSampler.serialize());
-        WHITELIST.serialize(json);
-        BLACKLIST.serialize(json);
+        INCLUDE_LIST.serialize(json);
+        EXCLUDE_LIST.serialize(json);
         KEEP_SCALING_ON_CONVERSION.serialize(json);
         SCALING_OVERRIDES.serialize(json);
         return json;
     }
 
-    public static void deserialize(JsonObject json) {
-        WHITELIST.deserialize(json);
-        BLACKLIST.deserialize(json);
-        KEEP_SCALING_ON_CONVERSION.deserialize(json);
-        SCALING_OVERRIDES.deserialize(json);
+    public static void deserialize(JsonObject json, Runnable setMakeBackup) {
+        INCLUDE_LIST.deserialize(json, setMakeBackup);
+        EXCLUDE_LIST.deserialize(json, setMakeBackup);
+        KEEP_SCALING_ON_CONVERSION.deserialize(json, setMakeBackup);
+        SCALING_OVERRIDES.deserialize(json, setMakeBackup);
 
         // deserialize default sampler
         try {
-            JsonElement defaultSamplerElement = json.get(DEFAULT_SAMPLER_NAME);
+            JsonElement defaultSamplerElement = GsonHelper.getAsJsonObject(json, DEFAULT_SAMPLER_NAME);
             defaultSampler = ScalingSampler.deserializeSampler(defaultSamplerElement, DEFAULT_SAMPLER_NAME);
         } catch (Exception e) {
-            RandomMobSizes.LOGGER.error("Error while parsing '{}', using default value", DEFAULT_SAMPLER_NAME);
+            RandomMobSizes.LOGGER.error("Error while parsing '{}', using default value.", DEFAULT_SAMPLER_NAME);
+            RandomMobSizes.LOGGER.error(e.getMessage());
             defaultSampler = createDefaultSampler();
-            throw new ConfigParseException(e.getMessage());
+            setMakeBackup.run();
         }
     }
 
-    private static GaussianScalingSampler createDefaultSampler() {
-        return new GaussianScalingSampler(0.5F, 1.5F, AttributeScalingTypes.NORMAL, AttributeScalingTypes.NORMAL, AttributeScalingTypes.INVERSE);
+    private static ScalingSampler createDefaultSampler() {
+        return new StaticScalingSampler(1F, AttributeScalingTypes.NONE, AttributeScalingTypes.NONE, AttributeScalingTypes.NONE);
     }
 
     public static ScalingSampler getScalingSampler(EntityType<?> entityType) {
@@ -63,10 +65,10 @@ public final class RandomMobSizesConfig {
         if (override != null) {
             return override;
         }
-        if (BLACKLIST.get().contains(entityType)) {
+        if (EXCLUDE_LIST.get().contains(entityType)) {
             return null;
         }
-        if (WHITELIST.get().contains(entityType)) {
+        if (INCLUDE_LIST.get().contains(entityType)) {
             return defaultSampler;
         }
         return null;