+++ Common/src/main/java/com/tristankechlo/random_mob_sizes/mixin_helper/MobMixinAddon.java
@@ -1,4 +1,4 @@
-package com.tristankechlo.random_mob_sizes.mixin_access;
+package com.tristankechlo.random_mob_sizes.mixin_helper;
 
 import net.minecraft.network.syncher.EntityDataAccessor;
 
@@ -9,5 +9,13 @@ public interface MobMixinAddon {
     void setMobScaling$RandomMobSizes(float scale);
 
     EntityDataAccessor<Float> getTracker$RandomMobSizes(); //needed for accessing the tracker from EntityMixin
+
+    boolean shouldScaleLoot$RandomMobSizes();
+
+    void setShouldScaleLoot$RandomMobSizes(boolean shouldScale);
+
+    boolean shouldScaleXP$RandomMobSizes();
+
+    void setShouldScaleXP$RandomMobSizes(boolean shouldScale);
 
 }